+++ src/main/java/org/springframework/beans/context/annotation/ClassPathBeanDefinitionScanner.java
@@ -1,6 +1,7 @@
 package org.springframework.beans.context.annotation;
 
 import cn.hutool.core.util.StrUtil;
+import org.springframework.beans.factory.annotation.AutowiredAnnotationBeanPostProcessor;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.stereotype.Component;
@@ -18,6 +19,9 @@ import java.util.Set;
  * @project: sbs-small-spring
  */
 public class ClassPathBeanDefinitionScanner extends ClassPathScanningCandidateComponentProvider {
+
+  public static final String AUTOWIRED_ANNOTATION_PROCESSOR_BEAN_NAME =
+      "org.springframework.context.annotation.internalAutowiredAnnotationProcessor";
 
   private BeanDefinitionRegistry registry;
 
@@ -46,6 +50,11 @@ public class ClassPathBeanDefinitionScanner extends ClassPathScanningCandidateCo
         registry.registerBeanDefinition(beanName, candidate);
       }
     }
+
+    // 注册处理@Autowired和@Value注解的BeanPostProcessor
+    registry.registerBeanDefinition(
+        AUTOWIRED_ANNOTATION_PROCESSOR_BEAN_NAME,
+        new BeanDefinition(AutowiredAnnotationBeanPostProcessor.class));
   }
 
   /**