+++ src/main/java/org/springframework/beans/factory/annotation/AutowiredAnnotationBeanPostProcessor.java
@@ -1,0 +1,64 @@
+package org.springframework.beans.factory.annotation;
+
+import cn.hutool.core.bean.BeanUtil;
+import org.springframework.beans.BeansException;
+import org.springframework.beans.PropertyValues;
+import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.BeanFactoryAware;
+import org.springframework.beans.factory.ConfigurableListableBeanFactory;
+import org.springframework.beans.factory.config.InstantiationAwareBeanPostProcessor;
+
+import java.lang.reflect.Field;
+
+/**
+ * 处理@Autowired和@Value注解的BeanPostProcessor
+ *
+ * @author: DoubleW2w
+ * @date: 2024/11/10
+ * @project: sbs-small-spring
+ */
+public class AutowiredAnnotationBeanPostProcessor
+    implements InstantiationAwareBeanPostProcessor, BeanFactoryAware {
+
+  private ConfigurableListableBeanFactory beanFactory;
+
+  @Override
+  public void setBeanFactory(BeanFactory beanFactory) throws BeansException {
+    this.beanFactory = (ConfigurableListableBeanFactory) beanFactory;
+  }
+
+  @Override
+  public PropertyValues postProcessPropertyValues(PropertyValues pvs, Object bean, String beanName)
+      throws BeansException {
+    // 处理@Value注解
+    Class<?> clazz = bean.getClass();
+    Field[] fields = clazz.getDeclaredFields();
+    for (Field field : fields) {
+      Value valueAnnotation = field.getAnnotation(Value.class);
+      if (valueAnnotation != null) {
+        String value = valueAnnotation.value();
+        value = beanFactory.resolveEmbeddedValue(value);
+        BeanUtil.setFieldValue(bean, field.getName(), value);
+      }
+    }
+    // todo：处理@autowired注解
+    return pvs;
+  }
+
+  @Override
+  public Object postProcessBeforeInitialization(Object bean, String beanName)
+      throws BeansException {
+    return bean;
+  }
+
+  @Override
+  public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
+    return bean;
+  }
+
+  @Override
+  public Object postProcessBeforeInstantiation(Class<?> beanClass, String beanName)
+      throws BeansException {
+    return null;
+  }
+}