+++ src/main/java/org/springframework/beans/util/StringValueResolver.java
@@ -1,0 +1,10 @@
+package org.springframework.beans.util;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/11/10
+ * @project: sbs-small-spring
+ */
+public interface StringValueResolver {
+  String resolveStringValue(String strVal);
+}